Updated version matching bare metal portal version

+++ app/v01/controllers/session.controller.js
@@ -2,9 +2,12 @@ var models  = require('../models');
 var Sequelize = require("sequelize");
 const Op = Sequelize.Op;
 var fs = require('fs');
+var hash = require ('string-hash');
+const path = require('path')
 
 const multer = require('multer');
 var messageCtrl = require('../controllers/message.controller');
+var tagCtrl = require('../controllers/tag.controller');
 
 var config  = require('../../config/config');
 
@@ -13,8 +16,9 @@ var ffmpeg = require('fluent-ffmpeg');
 
 var socketCtrl = require('../socket/socket.controller');
 
-
-
+const getColors = require('get-image-colors')
+var sizeOf = require('image-size');
+var rgb2hex = require('rgb2hex');
 
 var storage = multer.diskStorage({
 	destination: function (req, file, cb) {
@@ -125,6 +129,139 @@ function create(req, res){
 
 }
 
+var storageImg = multer.diskStorage({
+	destination: function (req, file, cb) {
+	  cb(null, config.baseDir + '/public/i')
+	},
+	filename: function (req, file, cb) {
+        var userId = req.auth.userId;
+        var ext = path.extname(file.originalname); 
+
+	    cb(null, hash(userId) + file.fieldname + '-' + Date.now() + ext)
+	}
+  })
+
+var uploadImageMw = multer({ storage: storageImg })
+
+function getImageColors(resultFile){
+
+    var promise = new Promise(function(resolve, reject) {
+        
+        getColors(resultFile).then(colors => {
+            // `colors` is an array of color objects 
+            let me = colors; 
+            console.log(me);
+    
+            let primColor = colors[0]._rgb; 
+    
+    
+            let rgbStr = "rgb(" + primColor[0] + "," + primColor[1] + "," + primColor[2] + ")"
+            let hexColor = rgb2hex(rgbStr).hex;
+            console.log(hexColor);
+
+            resolve(hexColor);
+    
+    
+        });
+
+        
+      });
+    
+      return promise;
+
+}
+
+
+function getImageMetaData(resultFile){
+
+    var getMetaData = new Promise(
+        function (resolve, reject) {
+
+            var dimensions = sizeOf(resultFile);
+             console.log(dimensions.width, dimensions.height);
+
+             var resolution = {
+                width : dimensions.width, 
+                height: dimensions.height
+            };
+            resolve(resolution); 
+
+        }
+    );
+
+    return getMetaData;
+
+
+}
+
+/**
+ * Upload base64 encoded string to create a session from a picture
+ */
+function uploadImage(req, res){
+
+    console.log(" HIER MUSS NOCH ÜBERPRÜFT WERDEN, DASS NUR DER OWNER ITEMS EINSTELLEN KANN ODER VORHER EINE GENEHMIGUNG ERTEILT WRERDEN MUSS")
+
+    var imageSrc = req.body.src;
+    var imagePurchaseTags = JSON.parse(req.body.newTags);
+    var userId = req.auth.userId;
+    var resultFilename = req.file.filename;
+    var fileType = req.file.mimetype;
+
+    let resultFile = path.join(config.baseDir , '/public/i', resultFilename); 
+
+    let colorCode;
+    
+    getImageColors(resultFile)
+    .then(function(color){
+        colorCode = color;
+        return resultFile
+    })
+    .then(getImageMetaData)
+    .then(function (resolution) {
+                const session = models.tblsessions.build({
+                    collectionId : req.params.collectionId, 
+                    sessionItemPath : "/i/" + req.file.filename, 
+                    sessionItemType : fileType, 
+                    sessionThumbnailPath : "/i/" + req.file.filename, 
+                    width: resolution.width, 
+                    height: resolution.height, 
+                    primeColor : colorCode
+                  }).save()
+                  .then(resultingSession => {
+                    
+                    console.log("after save"); 
+
+                    res.json(resultingSession);
+
+                    let sessionId = resultingSession.sessionId; 
+                    
+                    (async () => {
+                        await messageCtrl.notifySessionCreate(userId, sessionId);
+                        
+                        if (imagePurchaseTags.length > 0){
+                            await tagCtrl.createTags(sessionId, imagePurchaseTags);
+                        }
+                        
+                    })();
+
+                    return true;
+            
+                  })
+                  .catch(error => {
+                    // Ooops, do some error-handling
+                    console.log(error); 
+                    res.send(500, error);
+                  })
+
+            })
+            .catch(function (error) {
+                console.log(error); 
+                res.send(500, error);
+            });
+    
+}
+
+
 
 function deleteSession(req, res){
 
@@ -146,12 +283,15 @@ function deleteSession(req, res){
 
                 try{
                     fs.unlinkSync(itemPath);
-                    fs.unlinkSync(sessionThumbnailPath);
+
+                    if (itemPath != sessionThumbnailPath){
+                        fs.unlinkSync(sessionThumbnailPath);
+                    }
 
                     res.json({"message" : "ok"});
 
                 }catch(err){
-                    res.send(401, "files could not be found")
+                    res.send(500, "files could not be found")
                 }
 
                 }, function(error) {
@@ -169,4 +309,4 @@ function deleteSession(req, res){
 
 }
 
-module.exports =   { list, create, uploadVideo, deleteSession };
+module.exports =   { list, create, uploadVideo, uploadImage, uploadImageMw, deleteSession };
